refactor: SRP를 위한 메소드 책임 분리
- 스터디 지원 승인과 스터디원 생성 로직을 분리
